requirements.txt
asgiref==3.2.10
autopep8==1.5.3
Django==3.0.8
django-autocomplete-light==3.5.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-selectable==1.3.0
djangorestframework==3.11.0
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
